refactor(gsd-extension): move merge logic from WorktreeResolver into Lifecycle
Slice 7 / step D of ADR-016. The four merge-related verbs that lived on
WorktreeResolver — `mergeAndExit`, `_mergeWorktreeMode`,
`_mergeBranchMode`, `mergeAndEnterNext`, plus the no-merge `exitMilestone`
body — are now first-class Lifecycle methods. Lifecycle.exitMilestone
implements the merge-or-no-merge dispatch directly (no `resolverFactory`).

Concretely:
- WorktreeLifecycleDeps gains the entry-only-vs-now-merge-too dep set:
  isInAutoWorktree, autoCommitCurrentBranch, autoWorktreeBranch,
  teardownAutoWorktree, mergeMilestoneToMain, getCurrentBranch,
  checkoutBranch, resolveMilestoneFile, readFileSync.
- Worktree-mode merge now drives the worktree → root projection through
  Projection.finalizeProjectionForMerge(scope) instead of the legacy
  deps.syncWorktreeStateBack path-string call. The Lifecycle → Projection
  edge ADR-016 mandates is now load-bearing for both enter and exit.
- WorktreeResolver.mergeAndExit / mergeAndEnterNext / exitMilestone
  collapse to thin delegators (~10 lines each) that build a Lifecycle
  with a freshly-constructed Projection and forward the call. The legacy
  ~500-line bodies and their UserNotifiedError / isSamePath helpers are
  deleted from the file. The Resolver class retires entirely in step E.
- buildLifecycle() in auto.ts drops the `() => buildResolver()` arg now
  that resolverFactory is gone.

Test surface:
- Five obsolete delegation-shape tests in worktree-lifecycle.test.ts
  (resolverFactory required, fakeResolver delegation parity) are
  deleted with a note pointing readers at the merge-mode tests in
  worktree-resolver.test.ts that exercise the same behaviour through
  the remaining Resolver wrapper.
- One assertion in worktree-resolver.test.ts ("syncWorktreeStateBack
  was called") is dropped because that injected dep is no longer how
  the projection runs; the rest of the test still covers the merge
  behaviour end-to-end.
