+++ src/resources/extensions/gsd/worktree-lifecycle.ts
@@ -17,20 +17,30 @@
  * a circular reference. Both classes share the body until the Resolver retires.
  */
 
+import { existsSync, unlinkSync } from "node:fs";
 import { randomUUID } from "node:crypto";
+import { join } from "node:path";
 
 import type { AutoSession } from "./auto/session.js";
 import { debugLog } from "./debug-logger.js";
 import { emitJournalEvent } from "./journal.js";
-import { emitWorktreeCreated } from "./worktree-telemetry.js";
-import { resolveWorktreeProjectRoot } from "./worktree-root.js";
+import { emitWorktreeCreated, emitWorktreeMerged } from "./worktree-telemetry.js";
+import {
+  resolveWorktreeProjectRoot,
+  normalizeWorktreePathForCompare,
+} from "./worktree-root.js";
 import {
   claimMilestoneLease,
   refreshMilestoneLease,
   releaseMilestoneLease,
 } from "./db/milestone-leases.js";
 import { MergeConflictError } from "./git-service.js";
-import type { WorktreeResolver } from "./worktree-resolver.js";
+import {
+  getCollapseCadence,
+  getMilestoneResquash,
+  resquashMilestoneOnMain,
+} from "./slice-cadence.js";
+import { loadEffectiveGSDPreferences } from "./preferences.js";
 import type { WorktreeStateProjection } from "./worktree-state-projection.js";
 import { createWorkspace, scopeMilestone } from "./workspace.js";
 
@@ -54,21 +64,90 @@ export interface NotifyCtx {
  * recursion retires; shrinking it now would force a parallel migration.
  */
 export interface WorktreeLifecycleDeps {
+  // ── Entry / branch-mode setup ────────────────────────────────────────
   enterAutoWorktree: (basePath: string, milestoneId: string) => string;
   createAutoWorktree: (basePath: string, milestoneId: string) => string;
   enterBranchModeForMilestone: (basePath: string, milestoneId: string) => void;
   getAutoWorktreePath: (basePath: string, milestoneId: string) => string | null;
   getIsolationMode: (basePath?: string) => "worktree" | "branch" | "none";
+
+  // ── Cache + git service rebuild ──────────────────────────────────────
   invalidateAllCaches: () => void;
   GitServiceImpl: new (basePath: string, gitConfig: unknown) => unknown;
   loadEffectiveGSDPreferences: () =>
     | { preferences?: { git?: Record<string, unknown> } }
     | undefined;
+
+  // ── State Projection Module (ADR-016 one-way edge) ───────────────────
   /**
    * State Projection Module called by Lifecycle on enter/exit transitions.
    * Per ADR-016 the dependency direction is one-way: Lifecycle → Projection.
    */
   worktreeProjection: WorktreeStateProjection;
+
+  // ── Exit / merge / teardown ──────────────────────────────────────────
+  isInAutoWorktree: (basePath: string) => boolean;
+  autoCommitCurrentBranch: (
+    basePath: string,
+    reason: string,
+    milestoneId: string,
+  ) => void;
+  autoWorktreeBranch: (milestoneId: string) => string;
+  teardownAutoWorktree: (
+    basePath: string,
+    milestoneId: string,
+    opts?: { preserveBranch?: boolean },
+  ) => void;
+  mergeMilestoneToMain: (
+    basePath: string,
+    milestoneId: string,
+    roadmapContent: string,
+  ) => { pushed: boolean; codeFilesChanged: boolean };
+  getCurrentBranch: (basePath: string) => string;
+  /**
+   * Force-checkout the named branch in `basePath`. Required by the branch-mode
+   * merge path when HEAD has been moved off the milestone branch — silently
+   * skipping the merge would strand the milestone's commits.
+   */
+  checkoutBranch: (basePath: string, branch: string) => void;
+  resolveMilestoneFile: (
+    basePath: string,
+    milestoneId: string,
+    fileType: string,
+  ) => string | null;
+  /**
+   * Roadmap file reader. Injected so unit tests can substitute fixture
+   * content without touching the filesystem; production wiring passes
+   * `node:fs.readFileSync`.
+   */
+  readFileSync: (path: string, encoding: string) => string;
+}
+
+/**
+ * Internal sentinel — thrown by `_mergeBranchMode` when it has already
+ * emitted a user-visible error. The outer `mergeAndExit` catches the type
+ * and skips its own warning toast to avoid duplicate notifications.
+ */
+class UserNotifiedError extends Error {
+  readonly cause?: unknown;
+
+  constructor(message: string, cause?: unknown) {
+    super(message);
+    this.name = "UserNotifiedError";
+    this.cause = cause;
+  }
+}
+
+/**
+ * Compare two paths for physical identity, tolerating trailing slashes,
+ * symlink differences, and case variations on case-insensitive volumes.
+ *
+ * Used in place of string `===` / `!==` wherever one operand may be
+ * realpath-normalised and the other may not be (e.g. raw caller-supplied
+ * basePath vs. realpath-normalised projectRoot).
+ */
+function isSamePathPhysical(a: string, b: string): boolean {
+  return normalizeWorktreePathForCompare(a) === normalizeWorktreePathForCompare(b);
 }
 
 export type EnterResult =
@@ -452,16 +531,10 @@ function rebuildGitService(
 export class WorktreeLifecycle {
   private readonly s: AutoSession;
   private readonly deps: WorktreeLifecycleDeps;
-  private readonly resolverFactory?: () => WorktreeResolver;
 
-  constructor(
-    s: AutoSession,
-    deps: WorktreeLifecycleDeps,
-    resolverFactory?: () => WorktreeResolver,
-  ) {
+  constructor(s: AutoSession, deps: WorktreeLifecycleDeps) {
     this.s = s;
     this.deps = deps;
-    this.resolverFactory = resolverFactory;
   }
 
   /**
@@ -488,38 +561,16 @@ export class WorktreeLifecycle {
    * Unexpected failures (filesystem, git permissions, etc.) are wrapped
    * as `{ ok: false, reason: "teardown-failed", cause }` so callers always
    * receive a discriminated union — no exceptions for any expected outcome.
-   *
-   * Issue #5586 ships this as a delegating wrapper around
-   * `WorktreeResolver.mergeAndExit`/`exitMilestone`. The full extraction
-   * (~500 lines of merge logic across `_mergeWorktreeMode` and
-   * `_mergeBranchMode`) happens in #5587 when `WorktreeResolver` retires.
-   * The delegating shape preserves caller migration without rewriting
-   * merge-conflict handling mid-flight.
-   *
-   * `codeFilesChanged` reflects the actual value returned by
-   * `WorktreeResolver.mergeAndExit`. When `#5587` moves the merge logic
-   * into this Module directly, the delegation layer is removed but the
-   * contract is unchanged.
    */
   exitMilestone(
     milestoneId: string,
     opts: { merge: boolean; preserveBranch?: boolean },
     ctx: NotifyCtx,
   ): ExitResult {
-    if (!this.resolverFactory) {
-      throw new Error(
-        "WorktreeLifecycle.exitMilestone requires a resolverFactory until #5587 retires WorktreeResolver",
-      );
-    }
-    const resolver = this.resolverFactory();
     if (opts.merge) {
       try {
-        const mergeResult = resolver.mergeAndExit(milestoneId, ctx);
-        return {
-          ok: true,
-          merged: mergeResult?.merged ?? true,
-          codeFilesChanged: mergeResult?.codeFilesChanged ?? false,
-        };
+        const merged = this._mergeAndExit(milestoneId, ctx);
+        return { ok: true, merged, codeFilesChanged: false };
       } catch (err) {
         if (err instanceof MergeConflictError) {
           return { ok: false, reason: "merge-conflict", cause: err };
@@ -528,12 +579,584 @@ export class WorktreeLifecycle {
       }
     }
     try {
-      resolver.exitMilestone(milestoneId, ctx, {
+      this._exitWithoutMerge(milestoneId, ctx, {
         preserveBranch: opts.preserveBranch,
       });
       return { ok: true, merged: false, codeFilesChanged: false };
     } catch (err) {
       return { ok: false, reason: "teardown-failed", cause: err };
+    }
+  }
+
+  /**
+   * Milestone transition: merge the current milestone, then enter the next
+   * one. Pattern used when the loop detects that the active milestone has
+   * changed (current completed, next is now active). Caller is responsible
+   * for re-deriving state between the merge and the enter.
+   */
+  mergeAndEnterNext(
+    currentMilestoneId: string,
+    nextMilestoneId: string,
+    ctx: NotifyCtx,
+  ): void {
+    debugLog("WorktreeLifecycle", {
+      action: "mergeAndEnterNext",
+      currentMilestoneId,
+      nextMilestoneId,
+    });
+    try {
+      this._mergeAndExit(currentMilestoneId, ctx);
+    } catch (err) {
+      if (err instanceof UserNotifiedError) throw err;
+      // _mergeAndExit emits a warning and restores state on failure during
+      // merge/cleanup. If it throws before recovery runs (e.g. validation,
+      // emitJournalEvent), basePath isn't restored — re-throw so we don't
+      // enter the next milestone with the current one unmerged.
+      const projectRoot = resolveWorktreeProjectRoot(
+        this.s.basePath,
+        this.s.originalBasePath,
+      );
+      if (this.s.basePath !== projectRoot) throw err;
+    }
+    _enterMilestoneCore(this.s, this.deps, nextMilestoneId, ctx);
+  }
+
+  // ── Private — exit without merge ─────────────────────────────────────
+
+  private _exitWithoutMerge(
+    milestoneId: string,
+    ctx: NotifyCtx,
+    opts: { preserveBranch?: boolean },
+  ): void {
+    validateMilestoneId(milestoneId);
+    if (!this.deps.isInAutoWorktree(this.s.basePath)) {
+      debugLog("WorktreeLifecycle", {
+        action: "exitMilestone",
+        milestoneId,
+        skipped: true,
+        reason: "not-in-worktree",
+      });
+      return;
+    }
+
+    debugLog("WorktreeLifecycle", {
+      action: "exitMilestone",
+      milestoneId,
+      basePath: this.s.basePath,
+    });
+
+    try {
+      this.deps.autoCommitCurrentBranch(this.s.basePath, "stop", milestoneId);
+    } catch (err) {
+      debugLog("WorktreeLifecycle", {
+        action: "exitMilestone",
+        milestoneId,
+        phase: "auto-commit-failed",
+        error: err instanceof Error ? err.message : String(err),
+      });
+      ctx.notify(
+        `Auto-commit before exiting ${milestoneId} failed: ${err instanceof Error ? err.message : String(err)}. Branch ${this.deps.autoWorktreeBranch(milestoneId)} is preserved for recovery.`,
+        "warning",
+      );
+    }
+
+    if (this.s.originalBasePath) {
+      try {
+        process.chdir(this.s.originalBasePath);
+      } catch (err) {
+        debugLog("WorktreeLifecycle", {
+          action: "exitMilestone",
+          milestoneId,
+          phase: "pre-teardown-chdir-failed",
+          originalBasePath: this.s.originalBasePath,
+          error: err instanceof Error ? err.message : String(err),
+        });
+        ctx.notify(
+          `Could not leave milestone worktree before cleanup: ${err instanceof Error ? err.message : String(err)}. Branch ${this.deps.autoWorktreeBranch(milestoneId)} is preserved for recovery.`,
+          "warning",
+        );
+      }
+    }
+
+    let teardownFailed = false;
+    try {
+      this.deps.teardownAutoWorktree(this.s.originalBasePath, milestoneId, {
+        preserveBranch: opts.preserveBranch ?? false,
+      });
+    } catch (err) {
+      teardownFailed = true;
+      debugLog("WorktreeLifecycle", {
+        action: "exitMilestone",
+        milestoneId,
+        phase: "teardown-failed",
+        error: err instanceof Error ? err.message : String(err),
+      });
+      ctx.notify(
+        `Worktree cleanup failed for ${milestoneId}: ${err instanceof Error ? err.message : String(err)}. Branch ${this.deps.autoWorktreeBranch(milestoneId)} is preserved for recovery.`,
+        "warning",
+      );
+    }
+
+    this.restoreToProjectRoot();
+    debugLog("WorktreeLifecycle", {
+      action: "exitMilestone",
+      milestoneId,
+      result: "done",
+      basePath: this.s.basePath,
+    });
+    ctx.notify(
+      teardownFailed
+        ? `Worktree exit for ${milestoneId} needs manual cleanup.`
+        : `Exited worktree for ${milestoneId}`,
+      teardownFailed ? "warning" : "info",
+    );
+  }
+
+  // ── Private — merge and exit (worktree-mode or branch-mode) ──────────
+
+  /**
+   * Merge the completed milestone branch back to main and exit the worktree.
+   *
+   * - **worktree mode**: reads the roadmap, runs squash merge, projects
+   *   final state back via Projection.finalizeProjectionForMerge, tears
+   *   down the worktree, restores `s.basePath`. Falls back to bare
+   *   teardown (preserving the branch) if no roadmap exists.
+   * - **branch mode**: validates HEAD is on the milestone branch (recovers
+   *   via checkout if not), merges, rebuilds GitService.
+   * - **none**: no-op unless physically inside an auto-worktree (#2625).
+   *
+   * Returns true when an actual squash-merge ran. Throws MergeConflictError
+   * (and other non-recoverable errors) for callers to handle.
+   */
+  private _mergeAndExit(milestoneId: string, ctx: NotifyCtx): boolean {
+    validateMilestoneId(milestoneId);
+
+    // Anchor cwd at the project root before any merge work. Some merge
+    // paths (mergeMilestoneToMain, slice-cadence) chdir explicitly; others
+    // (branch-mode, isolation-degraded skip) do not. If the worktree dir
+    // is later torn down while cwd still points into it, every subsequent
+    // process.cwd() throws ENOENT — which after de73fb43d surfaces as a
+    // session-failed cancel and (in headless mode) terminates the whole
+    // gsd process. Best-effort: silent on failure so synthetic test paths
+    // still pass.
+    if (this.s.originalBasePath) {
+      try {
+        process.chdir(this.s.originalBasePath);
+      } catch (err) {
+        debugLog("WorktreeLifecycle", {
+          action: "mergeAndExit",
+          phase: "pre-merge-chdir-failed",
+          milestoneId,
+          originalBasePath: this.s.originalBasePath,
+          error: err instanceof Error ? err.message : String(err),
+        });
+      }
+    }
+
+    // #4764 — telemetry: record start timestamp so we can emit merge duration.
+    const mergeStartedAt = new Date().toISOString();
+    const mergeStartMs = Date.now();
+
+    if (this.s.isolationDegraded) {
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        milestoneId,
+        skipped: true,
+        reason: "isolation-degraded",
+      });
+      ctx.notify(
+        `Skipping worktree merge for ${milestoneId} — isolation was degraded (worktree creation failed earlier). Work is on the current branch.`,
+        "info",
+      );
+      return false;
+    }
+
+    const mode = this.deps.getIsolationMode(
+      this.s.originalBasePath || this.s.basePath,
+    );
+    debugLog("WorktreeLifecycle", {
+      action: "mergeAndExit",
+      milestoneId,
+      mode,
+      basePath: this.s.basePath,
+    });
+    emitJournalEvent(this.s.originalBasePath || this.s.basePath, {
+      ts: new Date().toISOString(),
+      flowId: randomUUID(),
+      seq: 0,
+      eventType: "worktree-merge-start",
+      data: { milestoneId, mode },
+    });
+
+    // #2625: If we are physically inside an auto-worktree, we MUST merge
+    // regardless of the current isolation config. This prevents data loss
+    // when the default isolation mode changes between versions.
+    const inWorktree =
+      this.deps.isInAutoWorktree(this.s.basePath) && this.s.originalBasePath;
+
+    if (mode === "none" && !inWorktree) {
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        milestoneId,
+        skipped: true,
+        reason: "mode-none",
+      });
+      return false;
+    }
+
+    let actuallyMerged = false;
+    if (mode === "worktree" || inWorktree) {
+      actuallyMerged = this._mergeWorktreeMode(milestoneId, ctx);
+    } else if (mode === "branch") {
+      actuallyMerged = this._mergeBranchMode(milestoneId, ctx);
+    }
+
+    if (!actuallyMerged) {
+      this.s.milestoneStartShas.delete(milestoneId);
+      return false;
+    }
+
+    // #4765 — when collapse_cadence=slice AND milestone_resquash=true, the
+    // N per-slice commits on main should be collapsed into one milestone
+    // commit. Done AFTER the primary merge-and-teardown so the branch and
+    // worktree are already cleaned up; we operate on main directly.
+    try {
+      const startSha = this.s.milestoneStartShas.get(milestoneId);
+      if (startSha) {
+        const prefs = loadEffectiveGSDPreferences(
+          this.s.originalBasePath || this.s.basePath,
+        )?.preferences;
+        if (
+          getCollapseCadence(prefs) === "slice" &&
+          getMilestoneResquash(prefs)
+        ) {
+          const result = resquashMilestoneOnMain(
+            this.s.originalBasePath || this.s.basePath,
+            milestoneId,
+            startSha,
+          );
+          if (result.resquashed) {
+            ctx.notify(
+              `slice-cadence: re-squashed slice commits for ${milestoneId} into a single milestone commit.`,
+              "info",
+            );
+          }
+        }
+        this.s.milestoneStartShas.delete(milestoneId);
+      }
+    } catch (err) {
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        milestoneId,
+        phase: "resquash",
+        error: err instanceof Error ? err.message : String(err),
+      });
+    }
+
+    // #4764 — record merge completion. Only reaches here when an actual
+    // merge ran; failure paths throw out before this point.
+    try {
+      emitWorktreeMerged(
+        this.s.originalBasePath || this.s.basePath,
+        milestoneId,
+        {
+          reason: "milestone-complete",
+          startedAt: mergeStartedAt,
+          durationMs: Date.now() - mergeStartMs,
+        },
+      );
+    } catch (telemetryErr) {
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        phase: "telemetry-emit",
+        error:
+          telemetryErr instanceof Error
+            ? telemetryErr.message
+            : String(telemetryErr),
+      });
+    }
+    return true;
+  }
+
+  /** Worktree-mode merge body. Returns true when an actual squash-merge ran. */
+  private _mergeWorktreeMode(milestoneId: string, ctx: NotifyCtx): boolean {
+    const originalBase = this.s.originalBasePath;
+    if (!originalBase) {
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        milestoneId,
+        mode: "worktree",
+        skipped: true,
+        reason: "missing-original-base",
+      });
+      return false;
+    }
+
+    let merged = false;
+    try {
+      // ADR-016: final projection before teardown. Replaces the legacy
+      // syncWorktreeStateBack(originalBase, basePath, milestoneId) call.
+      const finalScope = scopeMilestone(
+        createWorkspace(this.s.basePath),
+        milestoneId,
+      );
+      const { synced } =
+        this.deps.worktreeProjection.finalizeProjectionForMerge(finalScope);
+      if (synced.length > 0) {
+        debugLog("WorktreeLifecycle", {
+          action: "mergeAndExit",
+          milestoneId,
+          phase: "reverse-sync",
+          synced: synced.length,
+        });
+      }
+
+      // Resolve roadmap — try project root first, then worktree path as
+      // fallback. The worktree may hold the only copy when state-back
+      // projection silently dropped it or .gsd/ is not symlinked. Without
+      // the fallback, a missing roadmap triggers bare teardown which
+      // deletes the branch and orphans all milestone commits (#1573).
+      let roadmapPath = this.deps.resolveMilestoneFile(
+        originalBase,
+        milestoneId,
+        "ROADMAP",
+      );
+      if (
+        !roadmapPath &&
+        !isSamePathPhysical(this.s.basePath, originalBase)
+      ) {
+        roadmapPath = this.deps.resolveMilestoneFile(
+          this.s.basePath,
+          milestoneId,
+          "ROADMAP",
+        );
+        if (roadmapPath) {
+          debugLog("WorktreeLifecycle", {
+            action: "mergeAndExit",
+            milestoneId,
+            phase: "roadmap-fallback",
+            note: "resolved from worktree path",
+          });
+        }
+      }
+
+      if (roadmapPath) {
+        const roadmapContent = this.deps.readFileSync(roadmapPath, "utf-8");
+        const mergeResult = this.deps.mergeMilestoneToMain(
+          originalBase,
+          milestoneId,
+          roadmapContent,
+        );
+        merged = true;
+
+        // #2945 Bug 3: mergeMilestoneToMain performs best-effort worktree
+        // cleanup internally (step 12), but it can silently fail on Windows
+        // or when the worktree directory is locked. Perform a secondary
+        // teardown here to ensure the worktree is properly cleaned up.
+        // Idempotent — if already removed, teardownAutoWorktree no-ops.
+        try {
+          this.deps.teardownAutoWorktree(originalBase, milestoneId);
+        } catch {
+          // Best-effort — primary cleanup in mergeMilestoneToMain may have
+          // already removed the worktree.
+        }
+
+        if (mergeResult.codeFilesChanged) {
+          ctx.notify(
+            `Milestone ${milestoneId} merged to main.${mergeResult.pushed ? " Pushed to remote." : ""}`,
+            "info",
+          );
+        } else {
+          // #1906 — milestone produced only .gsd/ metadata. Surface
+          // clearly so the user knows the milestone is not truly complete.
+          ctx.notify(
+            `WARNING: Milestone ${milestoneId} merged to main but contained NO code changes — only .gsd/ metadata files. ` +
+              `The milestone summary may describe planned work that was never implemented. ` +
+              `Review the milestone output and re-run if code is missing.`,
+            "warning",
+          );
+        }
+      } else {
+        // No roadmap at either location — teardown but PRESERVE the branch
+        // so commits are not orphaned (#1573).
+        this.deps.teardownAutoWorktree(originalBase, milestoneId, {
+          preserveBranch: true,
+        });
+        ctx.notify(
+          `Exited worktree for ${milestoneId} (no roadmap found — branch preserved for manual merge).`,
+          "warning",
+        );
+      }
+    } catch (err) {
+      const msg = err instanceof Error ? err.message : String(err);
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        milestoneId,
+        result: "error",
+        error: msg,
+        fallback: "chdir-to-project-root",
+      });
+      emitJournalEvent(this.s.originalBasePath || this.s.basePath, {
+        ts: new Date().toISOString(),
+        flowId: randomUUID(),
+        seq: 0,
+        eventType: "worktree-merge-failed",
+        data: { milestoneId, error: msg },
+      });
+      // Surface a clear, actionable error. Worktree and milestone branch
+      // are intentionally preserved — nothing has been deleted. User can
+      // retry /gsd dispatch complete-milestone or merge manually once the
+      // underlying issue is fixed (#1668, #1891).
+      ctx.notify(
+        `Milestone merge failed: ${msg}. Your worktree and milestone branch are preserved — retry with \`/gsd dispatch complete-milestone\` or merge manually.`,
+        "warning",
+      );
+
+      // Clean up stale merge state left by failed squash-merge (#1389)
+      try {
+        const gitDir = join(originalBase || this.s.basePath, ".git");
+        for (const f of ["SQUASH_MSG", "MERGE_HEAD", "MERGE_MSG"]) {
+          const p = join(gitDir, f);
+          if (existsSync(p)) unlinkSync(p);
+        }
+      } catch {
+        /* best-effort */
+      }
+
+      // Error recovery: always restore to project root
+      if (originalBase) {
+        try {
+          process.chdir(originalBase);
+        } catch {
+          /* best-effort */
+        }
+      }
+
+      // Restore state before re-throwing so callers always get a
+      // consistent session (#4380).
+      this.restoreToProjectRoot();
+      // Re-throw: MergeConflictError stops the auto loop (#2330);
+      // non-conflict errors must also propagate so broken states are
+      // diagnosable (#4380).
+      throw err;
+    }
+
+    // Always restore basePath and rebuild — whether merge succeeded or failed
+    this.restoreToProjectRoot();
+    debugLog("WorktreeLifecycle", {
+      action: "mergeAndExit",
+      milestoneId,
+      result: "done",
+      basePath: this.s.basePath,
+    });
+    return merged;
+  }
+
+  /** Branch-mode merge body. Returns true when a merge actually ran. */
+  private _mergeBranchMode(milestoneId: string, ctx: NotifyCtx): boolean {
+    try {
+      const currentBranch = this.deps.getCurrentBranch(this.s.basePath);
+      const milestoneBranch = this.deps.autoWorktreeBranch(milestoneId);
+
+      if (currentBranch !== milestoneBranch) {
+        // #5538-followup: previous behaviour was to silently `return false`
+        // when HEAD wasn't on the milestone branch — that let the loop
+        // advance with the milestone's commits stranded on the branch.
+        // Attempt recovery by force-checking-out the milestone branch; if
+        // that fails, throw so the caller pauses auto-mode and the user
+        // sees the failure instead of a silent merge skip.
+        debugLog("WorktreeLifecycle", {
+          action: "mergeAndExit",
+          milestoneId,
+          mode: "branch",
+          recovery: "checkout-milestone-branch",
+          currentBranch,
+          milestoneBranch,
+        });
+        try {
+          this.deps.checkoutBranch(this.s.basePath, milestoneBranch);
+        } catch (checkoutErr) {
+          const checkoutMsg =
+            checkoutErr instanceof Error
+              ? checkoutErr.message
+              : String(checkoutErr);
+          ctx.notify(
+            `Cannot merge milestone ${milestoneId}: working tree is on ${currentBranch} and checkout to ${milestoneBranch} failed (${checkoutMsg}). Resolve manually and run /gsd auto to resume.`,
+            "error",
+          );
+          throw new UserNotifiedError(checkoutMsg, checkoutErr);
+        }
+
+        const reverify = this.deps.getCurrentBranch(this.s.basePath);
+        if (reverify !== milestoneBranch) {
+          const reverifyMsg = `branch checkout to ${milestoneBranch} reported success but current branch is ${reverify}`;
+          ctx.notify(
+            `Cannot merge milestone ${milestoneId}: ${reverifyMsg}. Resolve manually and run /gsd auto to resume.`,
+            "error",
+          );
+          throw new UserNotifiedError(reverifyMsg);
+        }
+      }
+
+      const roadmapPath = this.deps.resolveMilestoneFile(
+        this.s.basePath,
+        milestoneId,
+        "ROADMAP",
+      );
+      if (!roadmapPath) {
+        debugLog("WorktreeLifecycle", {
+          action: "mergeAndExit",
+          milestoneId,
+          mode: "branch",
+          skipped: true,
+          reason: "no-roadmap",
+        });
+        return false;
+      }
+
+      const roadmapContent = this.deps.readFileSync(roadmapPath, "utf-8");
+      const mergeResult = this.deps.mergeMilestoneToMain(
+        this.s.basePath,
+        milestoneId,
+        roadmapContent,
+      );
+
+      // Rebuild GitService after merge (branch HEAD changed)
+      rebuildGitService(this.s, this.deps);
+
+      if (mergeResult.codeFilesChanged) {
+        ctx.notify(
+          `Milestone ${milestoneId} merged (branch mode).${mergeResult.pushed ? " Pushed to remote." : ""}`,
+          "info",
+        );
+      } else {
+        ctx.notify(
+          `WARNING: Milestone ${milestoneId} merged (branch mode) but contained NO code changes — only .gsd/ metadata. ` +
+            `Review the milestone output and re-run if code is missing.`,
+          "warning",
+        );
+      }
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        milestoneId,
+        mode: "branch",
+        result: "success",
+      });
+      return true;
+    } catch (err) {
+      const msg = err instanceof Error ? err.message : String(err);
+      debugLog("WorktreeLifecycle", {
+        action: "mergeAndExit",
+        milestoneId,
+        mode: "branch",
+        result: "error",
+        error: msg,
+      });
+      if (!(err instanceof UserNotifiedError)) {
+        ctx.notify(`Milestone merge failed (branch mode): ${msg}`, "warning");
+      }
+      // Re-throw all errors so callers can apply their own recovery (#4380).
+      throw err;
     }
   }
